// lib/types/payload-types.ts
export interface Listing {
  id: string;
  photos: {
    photo: string | Photo;
    id?: string | null;
  }[];
  tag?: string | null;
  location: string;
  name: string;
  rooms: number;
  bathrooms: number;
  people: number;
  highlightedAmmenities: {
    highlightedAmmenity: string | AmmenityItem;
    id?: string | null;
  }[];
  description: string;
  listingHighlights?:
    | {
        icon: string | Icon;
        title: string;
        subtitle: string;
        id?: string | null;
      }[]
    | null;
  sleepingArrangement: {
    icons: {
      icon: string | Icon;
      id?: string | null;
    }[];
    bed: string;
    room: string;
    id?: string | null;
  }[];
  ammenityCards: {
    ammenityCardTitle: string;
    ammenityItems: {
      ammenityItem: string | AmmenityItem;
      id?: string | null;
    }[];
    id?: string | null;
  }[];
  availability: string;
  pricing: {
    periodName: string;
    period: string;
    price: number;
    priceType: "night" | "week";
    id?: string | null;
  }[];
  /**
   * @minItems 2
   * @maxItems 2
   */
  locationCoordinates: [number, number];
  locationItems: {
    icon: string | Icon;
    label: string;
    id?: string | null;
  }[];
  policies: {
    icon: string | Icon;
    text: string;
    id?: string | null;
  }[];
  updatedAt: string;
  createdAt: string;
  _status?: ("draft" | "published") | null;
}
export interface Photo {
  id: string;
  updatedAt: string;
  createdAt: string;
  url?: string | null;
  filename?: string | null;
  mimeType?: string | null;
  filesize?: number | null;
  width?: number | null;
  height?: number | null;
}
export interface AmmenityItem {
  id: string;
  icon: string | Icon;
  ammenity: string;
  updatedAt: string;
  createdAt: string;
}
export interface Icon {
  id: string;
  iconName: string;
  altAttribute: string;
  updatedAt: string;
  createdAt: string;
  url?: string | null;
  filename?: string | null;
  mimeType?: string | null;
  filesize?: number | null;
  width?: number | null;
  height?: number | null;
}